refactor(api): ajustado classes de entidade e adicionado migrations
