--- dev/backend/portallicitacao/src/main/java/com/ufs/es2/portallicitacao/model/Municipio.java
@@ -1,24 +1,0 @@
-package com.ufs.es2.portallicitacao.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Data
-public class Municipio {
-
-    @Id
-    private String codigoIBGE;
-    private String nome;
-    @ManyToOne
-    @JoinColumn(name = "estado_sigla")
-    private Estado estado;
-    private String regiao;
-    private String nomeIBGE;
-    private String codigoRegiao;
-    private String nomeRegiao;
-    private String pais;
-
-}
