+++ dev/backend/portallicitacao/src/main/java/com/ufs/es2/portallicitacao/models/Compra.java
@@ -1,0 +1,27 @@
+package com.ufs.es2.portallicitacao.model;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class Compra {
+
+    @Id
+    @Column(nullable = false, length = 45)
+    private String numero;
+    @Column(nullable = false, length = 45)
+    private String objeto;
+    @Column(nullable = false, length = 45)
+    private String numeroProcesso;
+    @Column(nullable = false, length = 45)
+    private String contatoResponsavel;
+    @OneToOne
+    @JoinColumn(name = "licitacao_idLicitacao")
+    private Licitacao licitacao;
+
+}
