+++ dev/backend/portallicitacao/src/main/java/com/ufs/es2/portallicitacao/models/Empenho.java
@@ -1,24 +1,32 @@
 package com.ufs.es2.portallicitacao.model;
 
 import jakarta.persistence.*;
-import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
-@Data
+@Getter
+@Setter
+@NoArgsConstructor
 public class Empenho {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    @Column(nullable = false, length = 45)
+    private String empenho;
+    @Column(nullable = false, length = 45)
     private String empenhoResumido;
+    @Column(nullable = false)
     private LocalDateTime dataEmissao;
+    @Column(nullable = false)
     private String observacao;
+    @Column(nullable = false)
     private BigDecimal valor;
     @ManyToOne
-    @JoinColumn(name = "licitacao_id")
+    @JoinColumn(name = "licitacao_idLicitacao")
     private Licitacao licitacao;
 
 }
