+++ dev/backend/portallicitacao/src/main/java/com/ufs/es2/portallicitacao/models/Fornecedor.java
@@ -1,0 +1,24 @@
+package com.ufs.es2.portallicitacao.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class Fornecedor {
+
+    @Id
+    @Column(nullable = false, length = 45)
+    private String cpfCnpj;
+    @Column(nullable = false, length = 45)
+    private String nome;
+    @Column(nullable = false, length = 45)
+    private String razaoSocial;
+
+}
