+++ dev/backend/portallicitacao/src/main/java/com/ufs/es2/portallicitacao/models/Licitacao.java
@@ -1,33 +1,44 @@
 package com.ufs.es2.portallicitacao.model;
 
 import jakarta.persistence.*;
-import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
-@Data
+@Getter
+@Setter
+@NoArgsConstructor
 public class Licitacao {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Integer idLicitacao;
+    @Column(nullable = false, length = 45)
     private String situacao;
+    @Column(nullable = false)
     private LocalDateTime dataResultadoCompra;
+    @Column(nullable = false)
     private LocalDateTime dataAbertura;
+    @Column(nullable = false)
     private LocalDateTime dataPublicacao;
+    @Column(nullable = false, length = 45)
     private String situacaoCompra;
+    @Column(nullable = false, length = 45)
     private String instrumentoLegal;
+    @Column(nullable = false)
     private BigDecimal valor;
     @ManyToOne
-    @JoinColumn(name = "municipio_codigo")
-    private String municipioCodigoIBGE;
+    @JoinColumn(name = "municipio_codigoIBGE")
+    private Municipio municipio;
     @ManyToOne
     @JoinColumn(name = "modalidade_codigo")
-    private String modalidadeCodigo;
+    private Modalidade modalidade;
     @ManyToOne
     @JoinColumn(name = "unidade_gestora_codigo")
-    private String unidadeGestora;
+    private UnidadeGestora unidadeGestora;
 
 }
