+++ dev/backend/portallicitacao/src/main/resources/application.properties
@@ -1,1 +1,3 @@
-
+spring.datasource.url=jdbc:mysql://localhost:3306/mydb
+spring.datasource.username=root
+spring.datasource.password=admin123
