+++ dev/backend/portallicitacao/src/main/resources/db/migration/V2__refatoracao.sql
@@ -1,0 +1,8 @@
+USE `mydb` ;
+
+ALTER TABLE `mydb`.`Licitacao`
+  DROP `situacai`;
+
+ALTER TABLE `mydb`.`Licitacao`
+  ADD `situacao` VARCHAR(45) NOT NULL;
+
